# build/swagger.yaml
openapi: 3.0.0
info:
    title: tempCodi-main
    contact: {}
    version: '1.0'
servers:
    - url: https://tempcodi-backend.onrender.com/api
      variables: {}
    - url: http://localhost:8280/api
      variables: {}
paths:
...
    /getCode:
        post:
            tags:
                - Misc
            summary: getWeatherCode
            operationId: getWeatherCode
            parameters: []
            requestBody:
                description: ''
                content:
                    application/json:
                        schema:
                            allOf:
                                - $ref: '#/components/schemas/getTempDataRequest'
                                - example:
                                      code: 500
                        example:
                            code: 500
                required: true
            responses:
                '200':
                    description: ''
                    headers: {}
            deprecated: false
            servers:
                - url: https://tempcodi-backend.onrender.com/api
                  variables: {}
...
components:
    schemas:
...
        getTempDataRequest:
            title: getTempDataRequest
            required:
                - lon
                - lat
                - city
            type: object
            properties:
                lon:
                    type: string
                lat:
                    type: string
                city:
                    type: string
            example:
                lon: '139.818482'
                lat: '35.618791'
                city: Koto
...